move spreaders to seperate files, improve fldpln algorithm
