+++ curve2flood/__init__.py
@@ -1,4 +1,4 @@
 from ._log import LOG
 from .core import Curve2Flood_MainFunction
 
-__all__ = ["Curve2Flood_MainFunction"]
+__all__ = ["Curve2Flood_MainFunction", "LOG"]
