+++ curve2flood/spreaders/__init__.py
@@ -1,0 +1,3 @@
+from .fldpln import fldpln
+from .kernel_weighted import create_kernel_weighted_spread_map
+from .multi_point import filter_outliers, compute_tw_multfact_scale, multi_point_interpolation
